fix: split oversized test files to meet 32KB limit
- Split call_graph.test.ts (52KB) into feature-based test files
- Split javascript.test.ts (41KB) into feature-based test files
- Created call_graph_extraction.test.ts for basic extraction tests
- Original files now contain pointers to the split files
- All files now under 32KB tree-sitter parsing limit
